requirements.txt
# Audio processing
sounddevice>=0.5.0
numpy>=2.0.0
scipy>=1.16.0
pyaudio>=0.2.14

# Speech-to-Text
faster-whisper>=1.1.0

# Text-to-Speech
piper-tts>=1.2.0

# Voice Activity Detection
torch>=2.0.0
torchaudio>=2.0.0
silero-vad>=4.0.0

# Wake word detection
pvporcupine>=3.0.0

# LLM Integration
anthropic>=0.40.0
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
